keep-n-tagged feature

+++ src/github-package.ts
@@ -54,22 +54,32 @@ export class GithubPackage {
     }
   }
 
-  async deletePackage(id: string, digest: string): Promise<void> {
-    core.info(`deleting package id: ${id} digest:${digest}`)
-    if (this.repoType === 'User') {
-      await this.config.octokit.rest.packages.deletePackageVersionForUser({
-        package_type: 'container',
-        package_name: this.config.name,
-        username: this.config.owner,
-        package_version_id: id
-      })
+  async deletePackage(
+    id: string,
+    digest: string,
+    tags: string[]
+  ): Promise<void> {
+    if (tags.length > 0) {
+      core.info(`deleting package id: ${id} digest:${digest} tag:${tags}`)
     } else {
-      await this.config.octokit.rest.packages.deletePackageVersionForOrg({
-        package_type: 'container',
-        package_name: this.config.name,
-        org: this.config.owner,
-        package_version_id: id
-      })
+      core.info(`deleting package id: ${id} digest:${digest}`)
+    }
+    if (!this.config.dryRun) {
+      if (this.repoType === 'User') {
+        await this.config.octokit.rest.packages.deletePackageVersionForUser({
+          package_type: 'container',
+          package_name: this.config.name,
+          username: this.config.owner,
+          package_version_id: id
+        })
+      } else {
+        await this.config.octokit.rest.packages.deletePackageVersionForOrg({
+          package_type: 'container',
+          package_name: this.config.name,
+          org: this.config.owner,
+          package_version_id: id
+        })
+      }
     }
   }
